pyproject.toml
[project]
name = "example-factory"
version = "0.1.0"
description = "Example Factory Application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12,<1",
    "sqlmodel>=0.0.24,<1",
    "python-slugify>=8.0.4,<9",
    "httpx>=0.28.1,<1",
    "pydantic>=2.11.4,<3",
    "pyyaml>=6.0.2,<7",
]

[tool.ruff]
exclude = [
    ".venv",
    "migrations",
]
line-length = 99

[tool.ruff.lint]
select = ["ALL"]

[tool.ruff.per-file-ignores]
"tests/*.py" = [
    "ANN001",  # Annotation-related warning
    "ANN201",  # Annotation-related warning
    "D100",  # Missing Docstring
    "D103",  # Missing Docstring
    "INP001",  # Add an '__init__.py'
    "PLR2004",  # Magic value
    "S101",  # Use of `assert`
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "hatch>=1.14.1",
    "hypothesis>=6.131.15",
    "mypy>=1.15.0",
    "pip-tools>=7.4.1,<8",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.8",
    "uv>=0.7.3",
    "uvicorn>=0.34.2",
    "pysocks>=1.7.1",
]

[tool.hatch.metadata]
license = "MIT"

[tool.hatch.build]
packages = ["."]
exclude = ["tests/"]
